refactor(parsing): move merchant types
