+++ src/parsing/merchants/index.ts
@@ -1,28 +1,10 @@
 import { boltzSwapMerchant } from "./boltz-swap-merchants"
 import { moneyBadgerMerchants } from "./money-badger-merchants"
+import type { Merchant, MerchantConfig } from "./types"
 import type { Network } from "../types"
 
 export { getIdentifierFromRegex } from "./helpers"
-
-export type MerchantCategory = "merchant-payment" | "swap"
-
-export type Merchant = {
-  id: string
-  lnurl: string
-  category: MerchantCategory
-  title: string
-  description: string
-  companyName: string
-  termsUrl: string
-  displayCurrency?: string
-}
-
-export type MerchantConfig = Omit<Merchant, "lnurl"> & {
-  getIdentifier: (input: string) => string | null
-  defaultDomain: string
-  domains: { [K in Network]: string }
-  getMerchants?: (identifier: string, network: Network) => Merchant[]
-}
+export type { Merchant, MerchantCategory, MerchantConfig } from "./types"
 
 export const merchants: MerchantConfig[] = [...moneyBadgerMerchants, boltzSwapMerchant]
